requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.2.2
frozenlist==1.3.1
future==0.18.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.3.0
MarkupSafe==2.1.1
multidict==6.0.2
requests==2.28.1
urllib3==1.26.11
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1
gunicorn==20.1.0
firebase-admin==5.3.0
pandas==1.5.0